requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
fastapi==0.95.2
ffmpy==0.3.0
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
gradio==3.32.0
gradio_client==0.2.5
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==2.0.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.3
openai==0.27.7
orjson==3.8.14
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pydantic==1.10.8
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.31.0
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.6.2
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.2
uvicorn==0.22.0
websockets==11.0.3
yarl==1.9.2
